Einige Ueberarbeitungen ...
git-svn-id: https://subversion.assembla.com/svn/oop-ws2012@70 8f0689cb-c7f0-493e-b951-6e7372201282

+++ OOP4/src/ClearBox.java
@@ -1,9 +1,9 @@
 public class ClearBox extends Box {
-	public ClearBox(int height, int width) {
+	public ClearBox(double height, double width) {
 		super(height, width, ' ', '*');
 	}
 	
 	public double getAspectRatio() {
-		return (double)width / height;
+		return width / height;
 	}
 }
